requirements.txt
ltlf2dfa
parse
logaut
pythomata
automata-lib
numpy
pyyaml
pm4py
loguru